requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
audioread==3.0.1
    # via librosa
certifi==2025.7.14
    # via requests
cffi==1.17.1
    # via soundfile
charset-normalizer==3.4.2
    # via requests
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via librosa
fonttools==4.59.0
    # via matplotlib
idna==3.10
    # via requests
joblib==1.5.1
    # via
    #   librosa
    #   scikit-learn
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via -r .\requirements.in
llvmlite==0.44.0
    # via numba
matplotlib==3.10.3
    # via -r .\requirements.in
msgpack==1.1.1
    # via librosa
numba==0.61.2
    # via librosa
numpy==2.2.6
    # via
    #   contourpy
    #   librosa
    #   matplotlib
    #   numba
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
packaging==25.0
    # via
    #   lazy-loader
    #   matplotlib
    #   pooch
pillow==11.3.0
    # via matplotlib
platformdirs==4.3.8
    # via pooch
pooch==1.8.2
    # via librosa
pyaudio==0.2.14
    # via -r .\requirements.in
pyaudiowpatch==0.2.12.7
    # via -r .\requirements.in
pycparser==2.22
    # via cffi
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
requests==2.32.4
    # via pooch
scikit-learn==1.7.1
    # via librosa
scipy==1.16.0
    # via
    #   librosa
    #   scikit-learn
six==1.17.0
    # via python-dateutil
soundfile==0.13.1
    # via librosa
soxr==0.5.0.post1
    # via librosa
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.14.1
    # via librosa
urllib3==2.5.0
    # via requests
websocket-client==1.8.0
    # via -r .\requirements.in